feat: rebuild website from scratch with Hexo
The current website was generated out of the archives of the previous one, built upon Wordpress.
Hopefully, the new website will make it far easier to improve the documentation.

The new website is build with Hexo (https://hexo.io/). The HTML content was converted to Markdown,
and the existing pages are preserved (only the blog is missing for now). The CSS style is more or
less the same as previously.

You may notice a bit of resemblance with the Vue.js (https://vuejs.org/) website. Let's hope we can
approach the level of quality of their documentation!