Update chat.md
very minor, but I didn't see a need for the "/" at the end of the <input> tag on the index.html.  But I'm a newbie, so if it's needed, maybe explain why? I deleted it on my local version of the sample chat, and it seems to work just the same.