Disconnection reasons updated
Added `transport close` reason and the existing reasons into a table with a description for each.

I believe that `transport close` reason is too important to add because disconnections in the major of cases are fired by network issues. I created this [condesandbox](https://codesandbox.io/s/socketio-playground-disconnect-event-6ojw9) like a playground to test this on server-side and client-side, with v3 in both but happen in v2 too.