attach connection listener before creating client
In the Testing examples (both v3 and v4) the client is created before the listener for connection events is created. Looks like this is no problem when using `socket.io@3` or `socket.io@4` together with the respective clients. However, it is a problem when using `socket.io@^3.1.x` together with a `socket.io-client@2` client having the `allowEIO3: true` option enabled on the server. Looks like the `"connection"` event passes before the listener is attached. Therefore, we should switch the order of these statements. These examples will continue to work on v3 and v4.